CORE-6153 Updated from review comments CORE-5791 removed destination from FlowSession KDocs
