+++ application/src/main/kotlin/net/corda/v5/application/marshalling/JsonMarshallingService.kt
@@ -4,6 +4,7 @@ import net.corda.v5.base.annotations.DoNotImplement
 
 /**
  * [JsonMarshallingService] marshalls to and from JSON using the registered JSON mapper.
+ * The platform will provide an instance of [JsonMarshallingService] to flows via property injection.
  */
 @DoNotImplement
 interface JsonMarshallingService : MarshallingService
