+++ application/src/main/kotlin/net/corda/v5/application/persistence/PagedQuery.kt
@@ -3,7 +3,7 @@ package net.corda.v5.application.persistence
 /**
  * Used to build a Query that supports limit and offset.
  *
- * @param R the type of the results.
+ * @param R The type of the results.
  */
 interface PagedQuery<R> : Query<R> {
     /**
@@ -11,9 +11,9 @@ interface PagedQuery<R> : Query<R> {
      * If no limit is set, all records will be returned.
      *
      * @param limit maximum number of results to return.
-     * @return the same [Query] instance.
+     * @return The same [PagedQuery] instance.
      *
-     * @throws IllegalArgumentException If [limit] is negative.
+     * @throws IllegalArgumentException if [limit] is negative.
      */
     fun setLimit(limit: Int): PagedQuery<R>
 
@@ -21,10 +21,10 @@ interface PagedQuery<R> : Query<R> {
      * Set the index of the first result in the query to return.
      * A default of `0` will be used in case it is not set.
      *
-     * @param offset the index of the first result in the query to return.
-     * @return the same [Query] instance.
+     * @param offset The index of the first result in the query to return.
+     * @return The same [PagedQuery] instance.
      *
-     * @throws IllegalArgumentException If [offset] is negative.
+     * @throws IllegalArgumentException if [offset] is negative.
      */
     fun setOffset(offset: Int): PagedQuery<R>
 }
