+++ application/src/main/kotlin/net/corda/v5/application/persistence/Query.kt
@@ -11,7 +11,7 @@ interface Query<R> {
     /**
      * Execute the [Query]
      *
-     * @return list of entities found. Empty list if none were found.
+     * @return List of entities found. Empty list if none were found.
      *
      * @throws CordaPersistenceException If there is an error executing the query.
      */
